fix(rabbitmq): Further cleanups after PR findings
