+++ tests/NetEvolve.HealthChecks.Tests.Unit/Azure/ServiceBus/ServiceBusTopicAdministrationTests.cs
@@ -2,7 +2,6 @@
 
 using System;
 using System.Diagnostics.CodeAnalysis;
-using System.IO;
 using System.Threading;
 using System.Threading.Tasks;
 using global::Azure.Messaging.ServiceBus.Administration;
